--- x-pack/plugin/async-search/qa/rest/build.gradle
@@ -1,32 +1,0 @@
-import org.elasticsearch.gradle.internal.info.BuildParams
-
-apply plugin: 'elasticsearch.base-internal-es-plugin'
-apply plugin: 'elasticsearch.internal-java-rest-test'
-apply plugin: 'elasticsearch.legacy-yaml-rest-test'
-apply plugin: 'elasticsearch.legacy-yaml-rest-compat-test'
-
-esplugin {
-  name = 'x-pack-test-deprecated-query'
-  description = 'Deprecated query plugin'
-  classname ='org.elasticsearch.query.DeprecatedQueryPlugin'
-}
-
-dependencies {
-  clusterPlugins project(xpackModule('async-search'))
-}
-
-restResources {
-  restApi {
-    include '_common', 'indices', 'index', 'async_search'
-  }
-}
-
-testClusters.configureEach {
-  testDistribution = 'DEFAULT'
-  setting 'xpack.security.enabled', 'false'
-}
-
-// Test clusters run with security disabled
-tasks.named("yamlRestTest") {
-  buildParams.withFipsEnabledOnly(it)
-}
