--- x-pack/plugin/async-search/qa/security/build.gradle
@@ -1,11 +1,0 @@
-apply plugin: 'elasticsearch.internal-java-rest-test'
-
-dependencies {
-  javaRestTestImplementation(testArtifact(project(xpackModule('core'))))
-  javaRestTestImplementation project(xpackModule('async-search'))
-  javaRestTestImplementation project(':test:framework')
-}
-
-tasks.named("javaRestTest").configure {
-  usesDefaultDistribution()
-}
