+++ x-pack/plugin/async-search/src/yamlRestTest/java/org/elasticsearch/qa/AsyncSearchRestIT.java
@@ -9,10 +9,18 @@ package org.elasticsearch.qa;
 
 import com.carrotsearch.randomizedtesting.annotations.ParametersFactory;
 
+import org.elasticsearch.test.cluster.ElasticsearchCluster;
 import org.elasticsearch.test.rest.yaml.ClientYamlTestCandidate;
 import org.elasticsearch.test.rest.yaml.ESClientYamlSuiteTestCase;
+import org.junit.ClassRule;
 
 public class AsyncSearchRestIT extends ESClientYamlSuiteTestCase {
+
+    @ClassRule
+    public static ElasticsearchCluster cluster = ElasticsearchCluster.local()
+        .module("x-pack-test-deprecated-query")
+        .module("x-pack-async-search")
+        .build();
 
     public AsyncSearchRestIT(final ClientYamlTestCandidate testCandidate) {
         super(testCandidate);
@@ -21,5 +29,10 @@ public class AsyncSearchRestIT extends ESClientYamlSuiteTestCase {
     @ParametersFactory
     public static Iterable<Object[]> parameters() throws Exception {
         return ESClientYamlSuiteTestCase.createParameters();
+    }
+
+    @Override
+    protected String getTestRestCluster() {
+        return cluster.getHttpAddresses();
     }
 }
