+++ x-pack/plugin/async-search/src/yamlRestTest/resources/rest-api-spec/test/async-search/10_deprecation.yml
