+++ x-pack/plugin/async-search/src/yamlRestTest/resources/rest-api-spec/test/async-search/20_keep_alive.yml
