+++ x-pack/test/deprecated-query/build.gradle
@@ -1,0 +1,14 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+apply plugin: 'elasticsearch.base-internal-es-plugin'
+
+esplugin {
+  name = 'x-pack-test-deprecated-query'
+  description = 'Deprecated query plugin'
+  classname ='org.elasticsearch.query.DeprecatedQueryPlugin'
+}
