+++ x-pack/test/deprecated-query/src/main/java/org/elasticsearch/query/DeprecatedQueryPlugin.java
